global_money_scanner/requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
ultralytics
python-multipart
python-dotenv
